Refactor log_pipeline into modular architecture with domain-specific packages
