+++ src/canonizer/__init__.py
@@ -1,0 +1,10 @@
+"""
+canonizer パッケージ
+
+データ正規化関連のモジュールを含む
+"""
+
+from .event import Event
+from .lol_canonizer import LoLCanonizer
+
+__all__ = ['Event', 'LoLCanonizer']
