+++ src/kpi/__init__.py
@@ -1,0 +1,11 @@
+"""
+kpi パッケージ
+
+KPI計算関連のモジュールを含む
+"""
+
+from .lol_kpi_config import LoLKPIConfig
+from .kpi_result import KPIResult
+from .lol_kpi_calculator import LoLKPICalculator
+
+__all__ = ['LoLKPIConfig', 'KPIResult', 'LoLKPICalculator']
