+++ src/llm/openrouter_client.py
@@ -1,0 +1,75 @@
+"""
+OpenRouter APIクライアントモジュール
+
+OpenRouter APIとの通信を管理する
+"""
+
+import json
+import logging
+from typing import Dict, Any
+
+
+class OpenRouterClient:
+    """OpenRouter APIクライアント"""
+    
+    def __init__(self, api_key: str = None, base_url: str = "https://openrouter.ai/api/v1"):
+        self.api_key = api_key or "dummy_key"  # テスト用
+        self.base_url = base_url
+        self.logger = logging.getLogger(__name__)
+        
+        # デフォルトモデル設定
+        self.primary_model = "anthropic/claude-3.5-sonnet"
+        self.fallback_models = [
+            "openai/gpt-4-turbo",
+            "openai/gpt-3.5-turbo",
+            "meta-llama/llama-3.1-70b-instruct"
+        ]
+        
+        # 使用統計
+        self.usage_stats = {
+            "requests": 0,
+            "total_tokens": 0,
+            "total_cost": 0.0,
+            "errors": 0
+        }
+    
+    async def request(self, prompt: str, model: str = None, max_tokens: int = 1000) -> Dict[str, Any]:
+        """OpenRouter APIリクエスト（最小実装）"""
+        model = model or self.primary_model
+        
+        # モックレスポンス（実際のAPI実装は後で追加）
+        mock_response = {
+            "choices": [{
+                "message": {
+                    "content": json.dumps({
+                        "analysis": {
+                            "performance_summary": f"Player analysis for {prompt[:50]}...",
+                            "key_strengths": ["テスト強み1", "テスト強み2"],
+                            "improvement_areas": ["テスト改善点1", "テスト改善点2"]
+                        },
+                        "recommendations": ["テスト推奨事項1", "テスト推奨事項2"],
+                        "champion_specific": {
+                            "role_analysis": "テストロール分析",
+                            "build_suggestions": "テストビルド提案",
+                            "positioning_tips": "テストポジション提案"
+                        }
+                    })
+                }
+            }],
+            "usage": {
+                "prompt_tokens": len(prompt.split()),
+                "completion_tokens": 100,
+                "total_tokens": len(prompt.split()) + 100
+            }
+        }
+        
+        # 統計更新
+        self.usage_stats["requests"] += 1
+        self.usage_stats["total_tokens"] += mock_response["usage"]["total_tokens"]
+        
+        self.logger.info(f"OpenRouter request completed: {model}, tokens: {mock_response['usage']['total_tokens']}")
+        return mock_response
+    
+    def get_usage_stats(self) -> Dict[str, Any]:
+        """使用統計を取得"""
+        return self.usage_stats.copy()
